Move contents of the core module into the "android.core" package
