+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/call/ZipCall.kt
@@ -2,7 +2,7 @@ package io.getstream.chat.android.client.call
 
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.utils.Result
-import io.getstream.chat.core.internal.coroutines.DispatcherProvider
+import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.async
