+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/core/internal/InternalStreamChatApi.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.core.internal
+package io.getstream.chat.android.core.internal
 
 @RequiresOptIn(
     message = "This is internal API for the Stream Chat libraries. Do not depend on this API in your own client code.",
