+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/core/internal/fsm/StateFunction.kt
@@ -1,3 +1,3 @@
-package io.getstream.chat.core.internal.fsm
+package io.getstream.chat.android.core.internal.fsm
 
 internal typealias StateFunction<S, E> = FiniteStateMachine<S, E>.(S, E) -> S
