+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/core/internal/fsm/builder/FSMBuilderMarker.kt
@@ -1,0 +1,7 @@
+package io.getstream.chat.android.core.internal.fsm.builder
+
+import io.getstream.chat.android.core.internal.InternalStreamChatApi
+
+@InternalStreamChatApi
+@DslMarker
+public annotation class FSMBuilderMarker
