--- stream-chat-android-core/src/main/java/io/getstream/chat/core/internal/fsm/builder/FSMBuilderMarker.kt
@@ -1,7 +1,0 @@
-package io.getstream.chat.core.internal.fsm.builder
-
-import io.getstream.chat.core.internal.InternalStreamChatApi
-
-@InternalStreamChatApi
-@DslMarker
-public annotation class FSMBuilderMarker
