+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/extensions/LiveDataExtensions.kt
@@ -1,7 +1,7 @@
 package io.getstream.chat.android.livedata.extensions
 
 import androidx.lifecycle.MutableLiveData
-import io.getstream.chat.core.internal.coroutines.DispatcherProvider
+import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import kotlinx.coroutines.withContext
 
 internal suspend fun <T : Any?> MutableLiveData<T>.setOnUi(newValue: T) =
