+++ stream-chat-android-test/build.gradle
@@ -16,7 +16,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
                 '-progressive',
                 '-Xexplicit-api=strict',
                 '-Xopt-in=kotlin.RequiresOptIn',
-                '-Xopt-in=io.getstream.chat.core.internal.InternalStreamChatApi',
+                '-Xopt-in=io.getstream.chat.android.core.internal.InternalStreamChatApi',
         ]
     }
 }
