+++ stream-chat-android-ui-components/build.gradle
@@ -67,7 +67,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
                 '-progressive',
                 '-Xexplicit-api=strict',
                 '-Xopt-in=kotlin.contracts.ExperimentalContracts',
-                '-Xopt-in=io.getstream.chat.core.internal.InternalStreamChatApi'
+                '-Xopt-in=io.getstream.chat.android.core.internal.InternalStreamChatApi'
         ]
     }
 }
