+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/utils/StartStopBuffer.kt
@@ -1,6 +1,6 @@
 package com.getstream.sdk.chat.utils
 
-import io.getstream.chat.core.internal.coroutines.DispatcherProvider
+import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
