+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/viewmodel/CreateChannelViewModel.kt
@@ -9,8 +9,8 @@ import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Member
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.models.name
+import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import io.getstream.chat.android.livedata.ChatDomain
-import io.getstream.chat.core.internal.coroutines.DispatcherProvider
 import kotlinx.coroutines.launch
 
 private val CHANNEL_NAME_REGEX = Regex("^!?[\\w-]*\$")
